Move path_lib from onnxruntime/core/framework to onnxruntime/core/platform.
